fix(omie-analytics-sync): rodada 2 do Codex — guard falha-total de inventory_position + coleta compartilhada (finitude/dedupe) no canônico e no full
Acatados do challenge xhigh (DO_NOT_SHIP → correções):
- [P1] falha TOTAL dos chunks de inventory_position agora ABORTA antes de
  estoque/custos (espelho do guard do sync-reprocess) — 'complete' com a fonte
  do cockpit/EOQ integralmente stale era o mesmo padrão do incidente original.
- [P2] syncInventoryFull coletava em ARRAY sem dedupe (código repetido no mesmo
  chunk = 21000) e sem validação de finitude — agora coleta via
  _shared/pos-estoque.ts (acumularPosicoesDaPagina movida de
  sync-reprocess/inventory-lote.ts com os testes, dedupe last-wins + descarte
  de item não-finito), synced_at único pós-coleta.
- [P2] syncInventory idem: normalização inline substituída pela função testada.

NÃO acatado (calibração registrada no PR): sondar páginas ALÉM do declarado —
reabriria decisão deliberada do #1341 (comportamento pós-fim do endpoint
desconhecido; fault viraria error permanente), compartilhada pelo sync-reprocess
em prod; sub-declaração inicial nunca observada no ListarPosEstoque.

Gate: deno test 119/119 · deno check 2 edges.
